large refactor of autoAssignOpenwater; create autoAssign module
